move read/write_items to item.c. fix bad memset in attrib.test.c.
